% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCSAL.R
\name{dcSAL}
\alias{dcSAL}
\title{contaminated shifted asymmetric Laplace density.}
\usage{
dcSAL(
  X,
  mu = rep(0, p),
  alpha = rep(0, p),
  sigma = diag(p),
  delta = 0.9,
  eta = 1.01
)
}
\arguments{
\item{mu}{A vector specifying the mean vector.}

\item{alpha}{A vector specifying the direction of each skewness in each variable.}

\item{sigma}{A matrix specifying the covariance matrix of the variables.}

\item{delta}{proportion of atypical values. 0 < delta < 1.}

\item{eta}{inflation parameter. eta > 1.}

\item{x}{A n by p matrix where each row corresponds a p-dimensional observation.}
}
\value{
A vector of length n that gives the value of the contaminated shifted asymmetric Laplace probability density function for each obervation in the matrix x and the specified parameter values.
}
\description{
Density function of the contaminated shifted asymmetric Laplace distribution.
}
