% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSAL.R
\name{dSAL}
\alias{dSAL}
\title{shifted asymmetric Laplace density.}
\usage{
dSAL(X, mu, alpha, sigma)
}
\arguments{
\item{mu}{A vector specifying the mean vector.}

\item{alpha}{A vector specifying the direction of each skewness in each variable.}

\item{sigma}{A matrix specifying the covariance matrix of the variables.}

\item{x}{A n by p matrix where each row corresponds a p-dimensional observation.}
}
\value{
A vector of length n that gives the value of the probability density function for each obervation in the matrix x and the specified parameter values.
}
\description{
Density function of the shifted asymmetri Laplace distribution.
}
